// src/app/core/models/masters/training/training-status.ts
export interface TrainingStatus {
    id: number;
    title: string;
    description: string;
    status: string;
    status_id :number;
    status_label: string;
}
export interface TrainingStatusPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: TrainingStatus[];
}

export interface TrainingStatusSaveResponse {
    id: number;
    message: string;
}